requirements.txt
aider-chat==0.86.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
cryptography==45.0.6
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.6
filelock==3.18.0
flake8==7.3.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grep-ast==0.9.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
litellm==1.75.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
mslex==1.3.0
multidict==6.6.3
networkx==3.4.2
numpy==1.26.4
openai==1.99.1
orjson==3.10.7
oslex==0.1.3
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
posthog==6.4.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
scipy==1.15.3
setuptools==80.9.0
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.41.3
tiktoken==0.10.0
tokenizers==0.21.4
tqdm==4.67.1
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.9.0
tree-sitter-yaml==0.7.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zipp==3.23.0